sre-python-utils/requirements.txt

typer>=0.12
rich>=13.7
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
tenacity>=8.2
apscheduler>=3.10
prometheus-api-client>=0.5
azure-identity>=1.16
